Switch order of arguments to prevent misinterpreting moves
