+++ implement/player.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 import numpy as np
-from move import Move
+from implement.move import Move
 
 
 class Player(ABC):
@@ -44,11 +44,11 @@ class HumanPlayer(Player):
 
     def get_move(self, game, legal_moves):
         move = self.decode_move(input(str(game)))
-        legal = move and move in legal_moves
+        legal = move and legal_moves[move]
         while not legal:
             print("Illegal move!")
             move = self.decode_move(input())
-            legal = move and move in legal_moves
+            legal = move and legal_moves[move]
         return move
 
     def receive_opp_move(self, move):
@@ -59,10 +59,12 @@ class HumanPlayer(Player):
         if len(move) != 3:
             return False
         char_to_num = {"A": 2, "C": 1, "B": 0}
+        # Place
         if move[0] in ["A", "B", "C"]:
             if not (move[1] in "0123" and move[2] in "0123"):
                 return False
-            return Move(True, int(move[1]), int(move[2]), ptype=char_to_num[move[0]])
+            return Move.encode_place(char_to_num[move[0]], int(move[1]), int(move[2]))
+        # Move
         else:
             if not (move[0] in "0123" and move[1] in "0123"):
                 return False
@@ -78,10 +80,9 @@ class HumanPlayer(Player):
                 dx = 1
             else:
                 return False
-            return Move(
-                False,
+            return Move.encode_move(
                 int(move[0]),
                 int(move[1]),
-                row2=int(move[0]) + dy,
-                col2=int(move[1]) + dx,
+                int(move[0]) + dy,
+                int(move[1]) + dx,
             )
